+++ tools/CopilotDesinger/copilot-designer/src/components/StageDetails.vue
@@ -2,7 +2,6 @@
   <h2>描述</h2>
   <textarea
     class="form-control"
-    id="details"
     placeholder="作业作者、参考的视频、危机合约要选哪些tag等等"
     v-model="modelValue"
   ></textarea>
